Add files via upload
This script checks the encoding of a data file to help the user determine whether the file is ready to import for data analysis. I created this after running into trouble with a CSV that would not read; turns out it was not UTF-8 encoded. This code helped me check the encoding so that I could change the file or get data from another source with properly encoded datasets.